:mushroom: Code simplification
