+++ components/furnituremix/frontend/src/app/core/components/search/index.tsx
@@ -1,13 +1,16 @@
 import { useEffect, useRef, useState } from 'react';
-import { Link } from 'react-router-dom';
 import SearchIcon from '~/assets/searchIcon.svg';
-import { useSearchInput } from '~/core/hooks/useSearchInput';
 import { DebounceInput } from 'react-debounce-input';
+import { useSuperFast } from 'src/lib/superfast/SuperFastProvider/Provider';
+import { createClient } from '@crystallize/js-api-client';
+import { Link } from '@remix-run/react';
+import { search } from '~/core/UseCases';
 
-export const SearchBar = ({}) => {
-    const { value, setValue, suggestions, setSuggestions, onChange } = useSearchInput('');
+export const SearchBar = () => {
     const ref = useRef<HTMLDivElement>(null);
     const [show, setShow] = useState(true);
+    const [suggestions, setSuggestions] = useState<any[]>([]);
+    const { state: superFastState } = useSuperFast();
 
     //close dropdown on outside click
     useEffect(() => {
@@ -23,6 +26,16 @@ export const SearchBar = ({}) => {
         };
     }, [ref]);
 
+    const handleChange = async (event: React.ChangeEvent<HTMLInputElement>) => {
+        const value = event.target.value;
+        const client = createClient({ tenantIdentifier: superFastState.config.tenantIdentifier });
+        try {
+            setSuggestions(await search(client, value));
+        } catch (error) {
+            console.error(error);
+        }
+    };
+
     return (
         <div className="bg-grey w-60 p-2 relative">
             <div className="flex gap-2">
@@ -30,19 +43,17 @@ export const SearchBar = ({}) => {
                 <DebounceInput
                     minLength={2}
                     debounceTimeout={300}
-                    defaultValue={value}
-                    onChange={onChange}
+                    onChange={handleChange}
                     className="bg-grey focus:border-textBlack outline-none"
                 />
             </div>
-            {suggestions?.length > 0 && show ? (
+            {suggestions.length > 0 && show ? (
                 <div ref={ref} className="absolute top-10 w-60 h-[200px] left-0 overflow-y-scroll shadow-sm">
                     {suggestions?.map((suggestion: any, index: number) => (
                         <div key={index}>
                             <Link
                                 to={suggestion?.node?.path}
                                 onClick={() => {
-                                    setValue('');
                                     setSuggestions([]);
                                 }}
                             >
